add new videos and version1 of final video
Version1 is first attempt at final video. It has the flow as we
discussed and placeholder slides of videos to add etc. It is saved in a
format that occupies least memory. later on i will save in better
resolution. It takes time to render thats why i saved in lesser memory
format